requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via httpx
async-timeout==4.0.3
    # via langchain
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via streamlit
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   tqdm
exceptiongroup==1.2.2
    # via anyio
faiss-cpu==1.10.0
    # via -r requirements.in
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.3.2
    # via
    #   huggingface-hub
    #   torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   langsmith
    #   ollama
huggingface-hub==0.30.1
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
    #   torch
joblib==1.4.2
    # via scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
langchain==0.3.23
    # via -r requirements.in
langchain-core==0.3.51
    # via
    #   langchain
    #   langchain-text-splitters
langchain-text-splitters==0.3.8
    # via langchain
langsmith==0.3.24
    # via
    #   langchain
    #   langchain-core
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
narwhals==1.33.0
    # via altair
networkx==3.2.1
    # via torch
numpy==2.0.2
    # via
    #   faiss-cpu
    #   pandas
    #   pydeck
    #   scikit-learn
    #   scipy
    #   streamlit
    #   transformers
ollama==0.4.7
    # via -r requirements.in
orjson==3.10.16
    # via langsmith
packaging==24.2
    # via
    #   altair
    #   faiss-cpu
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   streamlit
    #   transformers
pandas==2.2.3
    # via streamlit
pillow==11.1.0
    # via
    #   sentence-transformers
    #   streamlit
protobuf==5.29.4
    # via streamlit
pyarrow==19.0.1
    # via streamlit
pydantic==2.11.2
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   ollama
pydantic-core==2.33.1
    # via pydantic
pydeck==0.9.1
    # via streamlit
pymupdf==1.25.5
    # via -r requirements.in
pypdf2==3.0.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-core
    #   transformers
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   langchain
    #   langsmith
    #   requests-toolbelt
    #   streamlit
    #   transformers
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
safetensors==0.5.3
    # via transformers
scikit-learn==1.6.1
    # via sentence-transformers
scipy==1.13.1
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==4.0.2
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.40
    # via langchain
streamlit==1.44.1
    # via -r requirements.in
sympy==1.13.1
    # via torch
tenacity==9.1.2
    # via
    #   langchain-core
    #   streamlit
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.1
    # via transformers
toml==0.10.2
    # via streamlit
torch==2.6.0
    # via sentence-transformers
tornado==6.4.2
    # via streamlit
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.51.0
    # via sentence-transformers
typing-extensions==4.13.1
    # via
    #   altair
    #   anyio
    #   huggingface-hub
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   pypdf2
    #   referencing
    #   sentence-transformers
    #   sqlalchemy
    #   streamlit
    #   torch
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.3.0
    # via requests
watchdog==6.0.0
    # via streamlit
zstandard==0.23.0
    # via langsmith
